lxd: Add OIDC verifier
This adds an OIDC verifier which is used to verify access tokens
offline. It stores the issuer's remote keys so that they don't need to be
fetched every time a token needs to be validated. Also, this way there
is no need to call the issuer's token endpoint when a token needs to be
verified.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>